M3UA: Use %zu to print sizeof values

+++ print-m3ua.c
@@ -233,7 +233,7 @@ tag_value_print(netdissect_options *ndo,
     break;
   /* ... */
   default:
-    ND_PRINT("(length %u)", size + (u_int)sizeof(struct m3ua_param_header));
+    ND_PRINT("(length %zu)", size + sizeof(struct m3ua_param_header));
   }
   ND_TCHECK_LEN(buf, size);
   return;
